STEPPING IN PLACE WITH LEGS WORKING (FORCES STILL THROUGH GAZEBO)
This is a working commit, meaning stepping in place was succssfully tested, albeit with foot forces still being applied directly through gazebo, not the legs. Also, getting it to step in place is mostly luck it seems, which also has to be fixed.

I messed the commit on the walking_controller repo up, but I implemented a "pause sync", letting the walking controller know when the simulation is paused. Also, get_time() inside the walking controller is using the simulation time also sent via UDP to the walking controller.